Device list update on safe, EasyScan popup, Forcast fix for editing device ...

+++ app/src/main/java/com/example/savesomebills/Device.java
@@ -7,12 +7,12 @@ public class Device {
     public String groupId;
     public String name;
     public String icon;
-    public int onHours;
-    public int standbyHours;
+    public float onHours;
+    public float standbyHours;
     public int wattOn;
     public int wattStandby;
 
-    public Device(String name, String groupId, int onHours, int standbyHours, int wattOn, int wattStandby, String icon) {
+    public Device(String name, String groupId, float onHours, float standbyHours, int wattOn, int wattStandby, String icon) {
         this.objectId = UUID.randomUUID().toString();
         this.groupId = groupId;
         this.name = name;
